+++ gatsby-ssr.js
@@ -1,5 +1,5 @@
 import React from 'react';
 import { ThemeProvider } from 'emotion-theming';
-import { theme } from './src/theme/defaultTheme';
+import { theme } from './src/theme/docsTheme';
 
 export const wrapRootElement = ({ element }) => <ThemeProvider theme={theme}>{element}</ThemeProvider>;
